Add parsers module, add kbar to units
